renamed expanders package to amap
